战斗逻辑修改：添加换弹/子弹护盾等级

+++ Assets/Scripts/Game/Command/PlayerHealCommand.cs
@@ -6,9 +6,9 @@ using UnityEngine;
 /// </summary>
 public class PlayerHealCommand : AbstractCommand
 {
-    public int Amount { get; }
+    public float Amount { get; }
 
-    public PlayerHealCommand(int amount)
+    public PlayerHealCommand(float amount)
     {
         Amount = amount;
     }
